chore: Redesigned screen displayed when pressing "Lyrics' button on settings
- Now shows a list of the lyrics features.
- What was previously displayed when clicking the "Lyrics" button is now shown after clicking "Manage Lyrics".
