+++ mobile/src/modules/lyric/screens/ProviderView.tsx
@@ -6,7 +6,7 @@ import { View } from "react-native";
 
 import { Icon } from "~/resources/icons";
 import { useLyricStore } from "../core/store";
-import { moveLyricProvider, toggleCheckEmbeddedLyrics } from "../core/actions";
+import { moveLyricProvider } from "../core/actions";
 import type { LyricProvider } from "../core/constants";
 
 import { ContentPlaceholder } from "~/navigation/components/Placeholder";
@@ -17,9 +17,7 @@ import { Links, openLink } from "~/lib/web-browser";
 import { Button } from "~/components/Form/Button";
 import { FilledIconButton, IconButton } from "~/components/Form/Button/Icon";
 import { ListItem } from "~/components/List";
-import { SegmentedList } from "~/components/List/Segmented";
 import { StyledText } from "~/components/Typography/StyledText";
-import { Switch } from "~/components/UI/Switch";
 
 export default function LyricsProviders() {
   const { t } = useTranslation();
@@ -96,27 +94,18 @@ function RenderItem({ item, index }: DragListRenderItemInfo<LyricProvider>) {
 
 function Instructions() {
   const { t } = useTranslation();
-  const checkEmbeddedLyrics = useLyricStore((s) => s.checkEmbedded);
-
   return (
-    <View className="mb-6 gap-6">
-      <Button
-        onPress={() => openLink(Links.LyricsProviders)}
-        className="flex-row items-start pl-2"
-      >
-        <Icon name="info" size={20} color="onSurfaceVariant" />
-        <StyledText dim className="shrink grow text-sm">
-          {t("feat.lyrics.extra.providersInstructions.line1")}
-          {"\n\n"}
-          {t("feat.lyrics.extra.providersInstructions.line2")}
-        </StyledText>
-        <Icon name="open-in-new" />
-      </Button>
-      <SegmentedList.Item
-        labelTextKey="feat.lyrics.extra.useEmbedded"
-        onPress={toggleCheckEmbeddedLyrics}
-        RightElement={<Switch enabled={checkEmbeddedLyrics} />}
-      />
-    </View>
+    <Button
+      onPress={() => openLink(Links.LyricsProviders)}
+      className="mb-6 flex-row items-start pl-2"
+    >
+      <Icon name="info" size={20} color="onSurfaceVariant" />
+      <StyledText dim className="shrink grow text-sm">
+        {t("feat.lyrics.extra.providersInstructions.line1")}
+        {"\n\n"}
+        {t("feat.lyrics.extra.providersInstructions.line2")}
+      </StyledText>
+      <Icon name="open-in-new" />
+    </Button>
   );
 }
